add time of day interp and also fix zbuffer

+++ decompiler/level_extractor/BspHeader.h
@@ -131,7 +131,7 @@ struct TFragment : public Drawable {
   std::vector<u8> dma_base;
   // dma level 1 // 40 - 44
   std::vector<u8> dma_level1;
-  std::vector<u32> color_indices;  // note actually u64's
+  std::vector<u16> color_indices;
   u8 dma_qwc[4];
   // shader // 48 - 52
   u8 num_shaders;        // 52
